Adjust string converter and use it when calling string deserialization.
